# openapi.yaml
openapi: 3.0.3
info:
  title: PVZM Backend API
  description: "API for the Plants vs. Zombies: MODDED level sharing platform. Supports level uploading, downloading, browsing, favoriting, reporting, and admin management."
  version: 0.6.0
  contact:
    url: https://pvzm.net
servers:
  - url: https://backend.pvzm.net
    description: Production
...
paths:
...
  /api/admin/levels:
    get:
      tags: [Admin]
      summary: List levels (admin)
      description: Paginated level list with search. Requires GitHub OAuth.
      operationId: getAdminLevels
      security:
        - githubOAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: q
          in: query
          description: Search query (searches name, author, ID)
          schema:
            type: string
      responses:
        "200":
          description: Admin level listing
          content:
            application/json:
              schema:
                type: object
                required: [levels, total, page, limit, totalPages]
                properties:
                  levels:
                    type: array
                    items:
                      $ref: "#/components/schemas/LevelRecord"
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
                  totalPages:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/admin/levels/{id}:
    put:
      tags: [Admin]
      summary: Update a level
      description: Update level metadata. Requires GitHub OAuth or a valid one-time token.
      operationId: updateLevel
      security:
        - githubOAuth: []
        - oneTimeToken: []
      parameters:
        - $ref: "#/components/parameters/LevelId"
        - name: token
          in: query
          description: One-time admin token (alternative to OAuth)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                author:
                  type: string
                sun:
                  type: integer
                is_water:
                  type: integer
                  enum: [0, 1]
                difficulty:
                  type: integer
                favorites:
                  type: integer
                plays:
                  type: integer
                featured:
                  type: integer
                  enum: [0, 1]
                featured_at:
                  type: integer
                  nullable: true
      responses:
        "200":
          description: Level updated
          content:
            application/json:
              schema:
                type: object
                required: [success, level]
                properties:
                  success:
                    type: boolean
                    example: true
                  level:
                    $ref: "#/components/schemas/LevelRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Admin]
      summary: Delete a level
      description: Permanently deletes a level, its file, and all associated data. Requires GitHub OAuth or a valid one-time token.
      operationId: deleteLevel
      security:
        - githubOAuth: []
        - oneTimeToken: []
      parameters:
        - $ref: "#/components/parameters/LevelId"
        - name: token
          in: query
          description: One-time admin token (alternative to OAuth)
          schema:
            type: string
      responses:
        "200":
          description: Level deleted
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
                    example: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/levels/{id}/token:
    post:
      tags: [Admin]
      summary: Generate a one-time token for a level
      description: Creates a single-use token scoped to a specific level, allowing unauthenticated edit/delete access.
      operationId: generateToken
      security:
        - githubOAuth: []
      parameters:
        - $ref: "#/components/parameters/LevelId"
      responses:
        "200":
          description: Token generated
          content:
            application/json:
              schema:
                type: object
                required: [token, level_id]
                properties:
                  token:
                    type: string
                    example: token_abc123...
                  level_id:
                    type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/levels/{id}/feature:
    post:
      tags: [Admin]
      summary: Feature a level
      description: Marks a level as featured. Requires GitHub OAuth.
      operationId: featureLevel
      security:
        - githubOAuth: []
      parameters:
        - $ref: "#/components/parameters/LevelId"
      responses:
        "200":
          description: Level featured
          content:
            application/json:
              schema:
                type: object
                required: [success, level]
                properties:
                  success:
                    type: boolean
                    example: true
                  level:
                    $ref: "#/components/schemas/LevelRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Admin]
      summary: Unfeature a level
      description: Removes the featured status from a level. Requires GitHub OAuth.
      operationId: unfeatureLevel
      security:
        - githubOAuth: []
      parameters:
        - $ref: "#/components/parameters/LevelId"
      responses:
        "200":
          description: Level unfeatured
          content:
            application/json:
              schema:
                type: object
                required: [success, level]
                properties:
                  success:
                    type: boolean
                    example: true
                  level:
                    $ref: "#/components/schemas/LevelRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    githubOAuth:
      type: http
      scheme: bearer
      description: GitHub OAuth2 session-based authentication (via browser cookies)
    oneTimeToken:
      type: apiKey
      in: query
      name: token
      description: Single-use token scoped to a specific level ID
  parameters:
    LevelId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Level ID
  schemas:
...
    LevelRecord:
      type: object
      required: [id, name, author, sun, is_water, difficulty, favorites, plays, version, featured, featured_at, logging_data]
      properties:
        id:
          type: integer
        name:
          type: string
        author:
          type: string
        sun:
          type: integer
        is_water:
          type: integer
          enum: [0, 1]
        difficulty:
          type: integer
        favorites:
          type: integer
        plays:
          type: integer
        version:
          type: integer
        featured:
          type: integer
          enum: [0, 1]
        featured_at:
          type: integer
          nullable: true
        logging_data:
          type: string
          nullable: true
          description: JSON string containing Discord/Bluesky message IDs for logging management
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        retryAfterSeconds:
          type: number
          description: Present on 429 responses
  responses:
    BadRequest:
      description: Invalid input or validation failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...